[TASK] Change file names to UpperCamelCase
All files (except images) and folders must be renamed to use UpperCamelCase.

Some files and folders are renamed to be more explicit as well. Some parts
are ignored as they need to be restructured anyway. Images will have to be
given meaningful names over time, so it's pointless to transform them to
UpperCamelCase for now.

The references inside reST files are modified accordingly.

Sphinx will build this version without warnings and without errors.

Change-Id: Ic80404679de2078b5e8be7816e61c709fa622282