[TASK] Fix description of label_userFunc
The description of label_userFunc is fixed to mention that its
result is put through htmlspecialchars(), so it may not contain
any HTML markup.

Furthermore the forgotten label_userFunc_options property is
added.

The description of formattedLabel_userFunc property is also
modified to mention that it is used for the labels of inline
records only.

Resolves: #51413
Releases: 6.2

Change-Id: If4fb71faaa45f67424504582ec2df2fe873abe0b
Reviewed-on: https://review.typo3.org/27740
Reviewed-by: Francois Suter
Tested-by: Francois Suter