[BUGFIX] Remove broken palette "newline" documentation
The "--newline--" setting for part 3 was broken ever since. It was also documented wrong,
the correct setting would have been "newline" without the dashes. The underlying parsing
code dates back to Kasper's initial revision and was never touched or modified. The
according parsing code will be removed with CMS 7 and this documentation part can be
removed without substitution - also in 6.2 since it never worked anyway.