Update Index.rst
There is no real restriction for only 10 checkboxes as the underlying data integer (normaly int(11) in mysql) can hold 32 bit. as in reality the usage of the highest bit results in a loss of all bits, the maximum of possible checkboxes is 31.
I tested with bigint but all bits above 32 were ignored and setting bit 32 loses all bits, so I suspect a PHP restriction here.
Maybe a PHP expert who knows the differences about all php-implementations can validate if all current PHP versions work with 4Byte integers and so 31 bits are available everywhere.