Split up types with several renderTypes
The select type can have 5 different renderTypes. For each renderType a
number of properties exist and are listed.

This used to be all on one page which made it difficult to determine
which properties exist for one rendertype and when scrolling it could
easily happen that you were looking at a property for a different renderType.

This patch makes the following changes:

- Put properties and examples for each renderType on a subpage. This
  applies to the check, input, select and text type.
- Add a table of contents of all properties at the beginning of the
  properties section
- Change the ordering in the examples to list the code directly under
  the screenshot and not all screenshots and then all codeblocks (as already
  done with the check page).
- unify the header unlining
- put the headers in the main .rst and not in the included .rst.txt. This
  makes it easier to keep track of the hierarchy of header underlining.
  This also makes it not be a problem if a file is included in a different
  header level.
- leave a newline between header label and header. If this does not exist,
  cross-referencing does not work

Resolves: #258
Related: #253
Releases: master, 10.4, 9.5