[BUGFIX] Fix inverted description (#420)
* [BUGFIX] Fix inverted description

The TCA option localizeReferencesAtParentLocalization only works if
the field is NOT stored in a MM table.

Also add the select field as a possible type, as it works with all
fields defining foreign_table.

* [BUGFIX] Fix inverted description

The TCA option localizeReferencesAtParentLocalization only works if
the field is NOT stored in a MM table.

Also add the select field as a possible type, as it works with all
fields defining foreign_table.

* Minor change to wording

Co-authored-by: Tom Warwick <tom.warwick@typo3.org>
Co-authored-by: Nikita Hovratov <nikita.h@live.de>