[TASK] Move tableWizard explanation directly to textTable (#531)
The tableWizard is no longer a configurable option, but rather a
built-in feature, which cannot be changed. Thus, all options for
tableWizard are gone without replacements.

Unlike before, it is now displayed inline, instead of in a popup window.

The Text21 example has been removed, as it is the same as Text17 now.

In addition, the option "cols" can't be configured anymore for
textTable.

Fixes: #515
Releases: main, 11.5
(cherry picked from commit ad35a12fa465d3c072f3d7c3e078e9759ae92179)