Replace class name in static string in eval
Use ::class with class name instead of static string in the description
of the property eval.

Additionally, "required" is removed from the string, since this has been
moved to a separate property since v12.

Some captions reflecting the file names are added to the code blocks.

For v12, we can also add a "use" to ext_localconf.php to define the
full class name and use the short class name below.

Resolves: #599
Resolves: #252