Show properties for renderTypes
The input render types usually have common properties, but these
are not listed on the renderType pages, there is not even a link
there, only on the parent page.

This is a patch specific to v11.5 to add a link to the common properties
page to the input renderType pages. This is handled differently in
the versions >= 11.5.

Resolves: #668