Clarify foreign_table_where
The previous explanation for foreign_table_where was slightly
misleading. The examples speak for themselves, but they are
further down on the page.

The WHERE-clause is effectively appended to the existing WHERE
clause created by TYPO3 using the default values and must
start with AND.

Also, additional clauses such as ORDER BY or GROUP BY may be
added after the WHERE-clause.

Resolves: #663