Add info about wrong icon usage
It's the third time now that I stumbled over the problem, that in create new record wizard the icon of my group is from a foreign extension. This time it was the icon of tx_scheduler, which has nothing to do with my extension. TYPO3 tries to find a loaded extension by the second part of the tablename. This is OK for all extensions without underscore. But with underscore the second part of table name does not match any extension key. That's why it comes to these side-effects. Easiest solution is to enter a localized title. In that case the extension key will be extracted from the LLL:EXT: path. Else you should set groupName to your extension key. That way an extension can be found and the correct group icon (ext icon) will be shown again